+++ src/WebAPI/ClientApp/src/components/Navigation/ServerDialog/Tabs/ServerConfigTabContent.vue
@@ -3,7 +3,7 @@
 		<v-simple-table class="section-table">
 			<tbody>
 				<tr>
-					<td style="width: 50%">{{ $t('components.server-dialog.tabs.server-config.download-speed-limit') }}:</td>
+					<td style="width: 40%">{{ $t('components.server-dialog.tabs.server-config.download-speed-limit') }}:</td>
 					<td><download-limit-input :plex-server-id="plexServer.id" /></td>
 				</tr>
 			</tbody>
@@ -14,10 +14,7 @@
 <script lang="ts">
 import { Component, Vue, Prop } from 'vue-property-decorator';
 import type { PlexServerDTO } from '@dto/mainApi';
-
-@Component<ServerConfigTabContent>({
-	components: {},
-})
+@Component<ServerConfigTabContent>({})
 export default class ServerConfigTabContent extends Vue {
 	@Prop({ required: true, type: Object as () => PlexServerDTO })
 	readonly plexServer!: PlexServerDTO;